+++ src/ultimate_notion/file.py
@@ -2,12 +2,15 @@
 
 from __future__ import annotations
 
-from typing import cast
+from abc import ABC, abstractmethod
+from typing import Any, TypeVar, cast
 from urllib.parse import urlparse
+from uuid import UUID
 
-from emoji import emojize, is_emoji
+from emoji import demojize, emojize, is_emoji
 
 from ultimate_notion.core import Wrapper, get_repr
+from ultimate_notion.errors import InvalidAPIUsageError
 from ultimate_notion.obj_api import objects as objs
 from ultimate_notion.rich_text import Text, html_img
 
@@ -79,8 +82,39 @@ class FileInfo(Wrapper[objs.FileObject], wraps=objs.FileObject):
         self.obj_ref.value.url = url
 
 
-class Emoji(Wrapper[objs.EmojiObject], str, wraps=objs.EmojiObject):
-    """Emoji object which behaves like str."""
+TO = TypeVar('TO', bound=objs.TypedObject)
+
+
+class EmojiBase(Wrapper[TO], str, ABC, wraps=objs.TypedObject):
+    """Base class for emoji objects, which behave like str."""
+
+    @property
+    @abstractmethod
+    def name(self) -> str:
+        """Return the name of the emoji."""
+        raise NotImplementedError
+
+    @property
+    def to_code(self) -> str:
+        """Represent the emoji as :shortcode:, e.g. :smile:"""
+        return f':{self.name}:'
+
+    def __hash__(self):
+        return hash(str(self))
+
+    def __eq__(self, other: Any) -> bool:
+        if isinstance(other, EmojiBase | str):
+            return str(self) == str(other)
+        else:
+            return NotImplemented
+
+    def _repr_html_(self) -> str:  # noqa: PLW3201
+        """Called by JupyterLab automatically to display this file."""
+        return str(self)
+
+
+class Emoji(EmojiBase[objs.EmojiObject], wraps=objs.EmojiObject):
+    """Unicode emoji object which behaves like str."""
 
     def __init__(self, emoji: str) -> None:
         if not is_emoji(emoji):
@@ -90,45 +124,57 @@ class Emoji(Wrapper[objs.EmojiObject], str, wraps=objs.EmojiObject):
             raise ValueError(msg)
         self.obj_ref = objs.EmojiObject.build(emoji)
 
+    @property
+    def name(self) -> str:
+        """Return the name of the emoji."""
+        return demojize(self.obj_ref.emoji).strip(':')
+
     def __repr__(self) -> str:
-        return get_repr(self)
+        # we return the unicode emoji to the user for convenience in REPL
+        return self.obj_ref.emoji
 
     def __str__(self) -> str:
         return self.obj_ref.emoji
 
-    def __eq__(self, other: object) -> bool:
-        match other:
-            case str():
-                return str(self) == other
-            case Emoji():
-                return str(self) == str(other)
-            case _:
-                return NotImplemented
 
-    def __hash__(self):
-        return hash(str(self))
+class CustomEmoji(EmojiBase[objs.CustomEmojiObject], wraps=objs.CustomEmojiObject):
+    """Custom emoji object which behaves like str."""
 
-    def _repr_html_(self) -> str:  # noqa: PLW3201
-        """Called by JupyterLab automatically to display this file."""
-        return str(self)
+    def __init__(self) -> None:
+        msg = 'To create a new custom emoji, go to `Settings` -> `Emoji` in Notion.'
+        raise InvalidAPIUsageError(msg)
 
-    def to_code(self) -> str:
-        """Represent the emoji as :shortcode:, e.g. :smile:"""
-        raise NotImplementedError
+    @property
+    def name(self) -> str:
+        """Return the name of this custom emoji."""
+        return self.obj_ref.custom_emoji.name
 
-    @classmethod
-    def from_code(cls, shortcode: str) -> Emoji:
-        """Create an Emoji object from a :shortcode:, e.g. :smile:"""
-        raise NotImplementedError
+    @property
+    def id(self) -> UUID:
+        """Return the ID of this custom emoji."""
+        return self.obj_ref.custom_emoji.id
+
+    @property
+    def url(self) -> str:
+        """Return the URL of this custom emoji."""
+        return self.obj_ref.custom_emoji.url
+
+    def __repr__(self) -> str:
+        return get_repr(self, desc=self.name)
+
+    def __str__(self) -> str:
+        return f':{self.name}:'
 
 
-def wrap_icon(icon_obj: objs.FileObject | objs.EmojiObject) -> FileInfo | Emoji:
+def wrap_icon(icon_obj: objs.FileObject | objs.EmojiObject | objs.CustomEmojiObject) -> FileInfo | CustomEmoji | Emoji:
     """Wrap the icon object into the corresponding class."""
     match icon_obj:
         case objs.ExternalFile():
             return FileInfo.wrap_obj_ref(icon_obj)
         case objs.EmojiObject():
             return Emoji.wrap_obj_ref(icon_obj)
+        case objs.CustomEmojiObject():
+            return CustomEmoji.wrap_obj_ref(icon_obj)
         case _:
             msg = f'unknown icon object of {type(icon_obj)}'
             raise RuntimeError(msg)